benchmarks/rag_data_ingestion/requirements.txt
boto3==1.34.131
botocore==1.34.131
pyyaml==6.0.1
pytz==2024.1
PyPDF2==3.0.1
shortuuid==1.0.11
faiss-cpu==1.7.4
langchain==0.2.12
langchain-community==0.2.10
langchain-aws==0.1.13
langchain_postgres==0.0.9
pypdf==3.17.0
urllib3
psycopg[binary,pool]
psycopg2-binary